requirements.txt
discord.py==1.3.4
aioitertools==0.7.0
pymongo==3.10.1
dnspython==1.16.0
praw==7.1.0
git+https://github.com/Rapptz/discord-ext-menus